refactor: 마이페이지 컴포넌트 폴더명 및 경로 수정
